Giving up on putting the image in the URL
We can't put the uploaded image in the URL because the data
URL is too long. Putting the file name doesn't work, and having a URL
would require doing a fetch, and also would require the original image
to be online.

So we're leaving it as it. Still store the parameters in the URL but the
image will always have to be uploaded